Commit 0,3 Passing data
Added a pokemonBack image
Added a card prefab

Adjusted PokemonAPI.cs:
Now the script can pass data to the children of the Card prefab.
E.g. One of the children is text. With the PokemonAPI script I can take a specific string from the JSON file create a new string to put the info, and finally take that string and connect it with the text object.