Commit 0.4 fetch data for multiple cards
Added 49 duplicates of the card prefab to have 50 cards in total.
Added a CardHolder object to keep all the cards in.
Added a ScrollRect as parent of CardHolder to be able to scroll through the cards.
Added a Deck gameobject to represnt the deck building system.
Added some Plugins.

Adjusted button:
button is located in the Deck gameobject and can be pressed to fetch information to all 50 cards. The values of the cards are completely randomized due to a randomNum integer, which uses random.range to generate random data for the cards.